# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation",
    "version": "1.0.0",
    "title": "SK Enocean Gateway"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "gets a list of all devices",
        "description": "",
        "operationId": "getDevices",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "creates a new device",
        "description": "",
        "operationId": "addDevices",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/devices/{id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "get a sinlge device by id",
        "description": "",
        "operationId": "getDeviceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of device object",
            "required": true,
            "type": "string"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "updates an existing device",
        "description": "",
        "operationId": "updateDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of device object",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "updates an existing device",
        "description": "",
        "operationId": "deleteDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of device object",
            "required": true,
            "type": "string"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "gets all stored messages",
        "description": "",
        "operationId": "getMessages",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "get a single message",
        "description": "",
        "operationId": "getSingleMEssage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the message to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Messages"
        ],
        "summary": "delete a single message",
        "description": "",
        "operationId": "deleteSingleMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the message to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "apikey",
      "in": "header"
    }
  },
  "definitions": {
    "Device": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "eep": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "manufacturer": {
          "type": "string"
        }
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "messageString": {
          "type": "string"
        },
        "device_id": {
          "type": "string"
        }
      }
    }
  }
}